requirements.txt
# HEAMLAgentic dependencies
# Install with: pip install -r requirements.txt

# ── LLM Providers ────────────────────────────────────────────────────────────
anthropic>=0.40.0
openai>=1.50.0
google-genai>=1.0.0

# ── Web Interface ─────────────────────────────────────────────────────────────
streamlit>=1.34.0
plotly>=5.22.0

# ── ML & AutoML ───────────────────────────────────────────────────────────────
scikit-learn>=1.5.0
xgboost>=2.1.0
lightgbm>=4.5.0
optuna>=3.6.0
shap>=0.46.0
joblib>=1.4.0

# ── Materials Science ─────────────────────────────────────────────────────────
matminer>=0.9.2
pymatgen>=2024.6.10
mp-api>=0.41.0

# ── Data Collection ───────────────────────────────────────────────────────────
arxiv>=2.1.0
aflow>=0.0.14
pymupdf>=1.24.0
requests>=2.32.0

# ── Data ──────────────────────────────────────────────────────────────────────
pandas>=2.2.0
pyarrow>=16.0.0
numpy>=1.26.0
scipy>=1.13.0

# ── Reports ───────────────────────────────────────────────────────────────────
jinja2>=3.1.4

# ── Config ────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
